# BackEnd/PHP/php-laravel-composer-eloquent-postgresql/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'PBShop PHP API'
  description: 'Laravel implementation of the PBShop public API.'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
    description: 'Local development server'
  -
    url: /api/v1
paths:
...
  /api/v1/categories:
    get:
      tags:
        - Categories
      summary: '카테고리 트리 조회'
      operationId: categoryIndex
      responses:
        200:
          description: '조회 성공'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
    post:
      tags:
        - Categories
      summary: '카테고리 생성'
      operationId: categoryStore
      responses:
        201:
          description: '생성 성공'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
  '/api/v1/categories/{id}':
    get:
      tags:
        - Categories
      summary: '카테고리 상세 조회'
      operationId: categoryShow
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: '상세 조회 성공'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
    delete:
      tags:
        - Categories
      summary: '카테고리 삭제'
      operationId: categoryDestroy
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: '삭제 성공'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
    patch:
      tags:
        - Categories
      summary: '카테고리 수정'
      operationId: categoryUpdate
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: '수정 성공'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
...
  '/api/v1/wishlist/{productId}':
    post:
      tags:
        - Wishlist
      summary: '위시리스트 추가 또는 토글'
      operationId: 2639744d97b31c7336f8f8ee45385aca
      parameters:
        -
          name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Wishlist
      summary: '위시리스트 삭제'
      operationId: f1e32d70d0c2378fc630eb8b025228c2
      parameters:
        -
          name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
...
  /api/v1/images/upload:
    post:
      tags:
        - Image
      summary: '이미지 업로드'
      operationId: f5e93b029ed8251cf67a14317c5c4d4f
      responses:
        201:
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
  '/api/v1/images/{id}/variants':
    get:
      tags:
        - Image
      summary: '이미지 변환본 조회'
      operationId: f8d63be7bfca6b6035d15f1b90b353eb
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
  '/api/v1/images/{id}':
    delete:
      tags:
        - Image
      summary: '이미지 삭제'
      operationId: 697898032dfad9f26f8209c42367e4fd
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ApiMeta:
      description: '공통 메타 정보'
      properties:
        requestId:
          type: string
          example: req_123456
          nullable: true
        locale:
          type: string
          example: ko
          nullable: true
        currency:
          type: string
          example: KRW
          nullable: true
        page:
          type: integer
          example: 1
          nullable: true
        limit:
          type: integer
          example: 20
          nullable: true
        totalCount:
          type: integer
          example: 100
          nullable: true
        totalPages:
          type: integer
          example: 5
          nullable: true
      type: object
      additionalProperties: true
...
    ApiSuccessEnvelope:
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          description: '엔드포인트별 실제 응답 데이터'
          nullable: true
          oneOf:
            -
              type: object
              additionalProperties: true
            -
              type: array
              items: []
            -
              type: string
            -
              type: integer
            -
              type: number
            -
              type: boolean
        meta:
          oneOf:
            -
              $ref: '#/components/schemas/ApiMeta'
          nullable: true
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Bearer token'
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []